slight changes

+++ lib/models/analysis/score_distribution_item.dart
@@ -2,5 +2,5 @@ class ScoreDistributionItem {
   const ScoreDistributionItem(this.label, this.value);
 
   final String label;
-  final double value;
+  final int value;
 }
